+++ clients/margin-trading/examples/rest-api/TradeApi/marginAccountNewOtoco.ts
@@ -13,10 +13,10 @@ async function marginAccountNewOtoco() {
             symbol: 'symbol_example',
             workingType: 'workingType_example',
             workingSide: 'workingSide_example',
-            workingPrice: 1,
-            workingQuantity: 1,
+            workingPrice: 1.0,
+            workingQuantity: 1.0,
             pendingSide: 'pendingSide_example',
-            pendingQuantity: 1,
+            pendingQuantity: 1.0,
             pendingAboveType: 'pendingAboveType_example',
         });