+++ clients/spot/src/websocket-api/types/exchange-info-response-result-symbols-inner.ts
@@ -130,7 +130,7 @@ export interface ExchangeInfoResponseResultSymbolsInner {
      * @type {boolean}
      * @memberof ExchangeInfoResponseResultSymbolsInner
      */
-    allowAmend?: boolean;
+    amendAllowed?: boolean;
     /**
      *
      * @type {boolean}