+++ clients/crypto-loan/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 6.0.3 - 2025-06-16
+
+### Changed (1)
+
+- Update `@binance/common` library to version `1.1.1`.
+
 ## 6.0.2 - 2025-06-05
 
 ### Changed (1)